Formatting, MongoDB testing.
As of right now, there are a few issues with the application.
The GET Requests are not recognized or even taken into consideration, I
do not know why, however, this causes an issue for all of the GET
requests that contain parameters.

However, somehow all of those that do not have parameters, seem to work.
Other than that, the MongoDB string used to find the results according
to user ID, is,
`db.post.find({"userID": ObjectId("--someObjId--")}).limit(n)`.